Adds envs and only offer 'theme' commands
Farewell theme-configure and single configuration files, this
moves `config.yml` into an environment based setup such that
you as a developer can easily run your commands on various
different environments quickly and easily.

Because of the change over to using environments for theme
management and to reduce the amount of overhead in tickets
that might rise due to the different commands behaving
differently, I've chosen to remove theme-<suffix> commands
from the releases.

If you have an old configuration, it will be migrated over
to the new layout automatically for you.

The environments are basic YAML files with each key containing
a configuration within it. They'll look something like this:

```yaml
development:
  access_token: abracadabra
  store: something-dev.myshopify.com
  bucket_size: 40
  refill_rate: 2
staging:
  access_token: alakazam
  store: something.myshopify.com
  theme_id: 1234
  bucket_size: 40
  refill_rate: 2
production:
  access_token: alakazam
  store: something.myshopify.com
  bucket_size: 40
  refill_rate: 2
```