Prototyping / Trying out some ways to improve logs
The library did a bunch of weird stuff like logging
to console, which isn't cool if we ever want to use
this like an actual library and not just as a command
line utility.

All this stuff is broken, console logging is missing and
so on. This will be addressed soonish