Fixes crash when configuration is invalid
There was a case where invalid YAML configuration files would
cause the program to crash. This also fixes an issue where an
invalid configuration file would cause the application to think
that the configuration simply is out dated, even though that isn't
the case.

Improves an error message such that people can get an idea of what
they could be doing wrong.