Fixes File Descriptor exhaustion
Files were not being closed after they were used.
Typically this didn't really cause any issues, but sometimes
when working on large projects or when running the application
for an extended period of time the app would crash with a
really undescriptive error message.

The underlying issue was because we were keeping files open.

By defering Close on the file we ensure that we don't exhaust our
File Descriptor limits.