Make install script work with Python 3
Current install script uses print statements without parentheses, which causes errors in Python 3...

$ python3 install
  File "install", line 75
    print "Downloading themekit from %s/%s" % (self.host, self.path)
                                          ^
SyntaxError: invalid syntax

For this new version of install, I used 2to3 to fix the print statements and a few other minor issues. It works with 2.7, 3.4 and probably any Python version in between. 

This does NOT work with 3.5 because of some breaking module name changes with httplib and urlparse.