Use try-catch to exit PowerShell script on error
Script should not continue and report "Install Complete" when there is an error. Instead, it should print out an error message and exit with an exit code indicating failure.

Note that Write-Output writes to the pipeline while Write-Host writes to the host terminal (usually the user's screen). Only Write-Host provides coloring options.